// src/app/data/generatedReels.ts
export const generatedProducts: Product[] = [
  {
    "id": "gp-aqmnhlwmydok4qhw-tm58aee-c4afcf",
    "category": "accessories",
    "name": "AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC...",
    "price": 406500000,
    "image": "https://picsum.photos/seed/product-aqmnhlwmydok4qhw-tm58aee-c4afcf/800/800",
    "description": "محصول AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4.3,
    "reviews": 663
  },
  {
    "id": "gp-aqmtsdupemne5rre9nl4qv0r-916755",
    "category": "wearables",
    "name": "AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ...",
    "price": 426500000,
    "image": "https://picsum.photos/seed/product-aqmtsdupemne5rre9nl4qv0r-916755/800/800",
    "description": "محصول AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4.3,
    "reviews": 613
  },
  {
    "id": "gp-aqnkeqmwe4y1x0y8exa0covz-416d31",
    "category": "home",
    "name": "AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg...",
    "price": 214000000,
    "image": "https://picsum.photos/seed/product-aqnkeqmwe4y1x0y8exa0covz-416d31/800/800",
    "description": "محصول AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4.8,
    "reviews": 588
  },
  {
    "id": "gp-aqnncqgdgprt6qbzz-q6vxwy-1cbe7e",
    "category": "fashion",
    "name": "AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj...",
    "price": 102000000,
    "image": "https://picsum.photos/seed/product-aqnncqgdgprt6qbzz-q6vxwy-1cbe7e/800/800",
    "description": "محصول AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4.4,
    "reviews": 614
  },
  {
    "id": "gp-aqonqljlvkpnzf35gv6bgc_l-f44949",
    "category": "beauty",
    "name": "AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ...",
    "price": 230000000,
    "image": "https://picsum.photos/seed/product-aqonqljlvkpnzf35gv6bgc_l-f44949/800/800",
    "description": "محصول AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4,
    "reviews": 490
  },
  {
    "id": "gp-aqozgdfofpqucrb0hhk-cscp-a77081",
    "category": "wearables",
    "name": "AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03...",
    "price": 336500000,
    "image": "https://picsum.photos/seed/product-aqozgdfofpqucrb0hhk-cscp-a77081/800/800",
    "description": "محصول AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4.3,
    "reviews": 653
  },
  {
    "id": "gp-aqpvipy3xh57cg-ueelcjdr8-e219a7",
    "category": "home",
    "name": "AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv...",
    "price": 589000000,
    "image": "https://picsum.photos/seed/product-aqpvipy3xh57cg-ueelcjdr8-e219a7/800/800",
    "description": "محصول AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv... با کیفیت مناسب و قیمت رقابتی برای بازار ایران.",
    "sizes": [
      "One Size"
    ],
    "colors": [
      "مشکی",
      "سفید",
      "نقره‌ای"
    ],
    "rating": 4.8,
    "reviews": 538
  }
];
// ...
export const generatedReelCommentsFa: Record<string, string[]> = {
  "gv-aqmnhlwmydok4qhw-tm58aee-c4afcf": [
    "این AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQMNHLwMyDok4Qhw TM58AEewKFOUhu B1D67wAgDAdjA0xsC... منطقیه 💸"
  ],
  "gv-aqmtsdupemne5rre9nl4qv0r-916755": [
    "این AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQMtsdUPEMne5RrE9Nl4qV0RO97iOEWmMOo9hcTQHPgsa7SkZ... منطقیه 💸"
  ],
  "gv-aqnkeqmwe4y1x0y8exa0covz-416d31": [
    "این AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQNKEQMwE4y1x0Y8ExA0COVZ mf4rDl gPJKyHzXBYImN CUg... منطقیه 💸"
  ],
  "gv-aqnncqgdgprt6qbzz-q6vxwy-1cbe7e": [
    "این AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQNnCQGDGPRt6QBZZ Q6vXwyWRpdWrD GDnHX jvtGv94sdwj... منطقیه 💸"
  ],
  "gv-aqonqljlvkpnzf35gv6bgc_l-f44949": [
    "این AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQOnQLjlvkpnZF35gv6BGC lLAxLKRAQpqSggae2sX5uOewy ... منطقیه 💸"
  ],
  "gv-aqozgdfofpqucrb0hhk-cscp-a77081": [
    "این AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQOzgDFofpQUCrb0hhK Cscpf2195 InXtdLeCEzCCNztPV03... منطقیه 💸"
  ],
  "gv-aqpvipy3xh57cg-ueelcjdr8-e219a7": [
    "این AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv... موجوده؟ قیمتش خیلی خوبه 👌",
    "کیفیت AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv... در چه حدیه؟ کسی خریده؟",
    "برای خرید AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv... ارسال فوری هم دارید؟",
    "رنگ‌بندی AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv... کامل‌تر هم هست؟",
    "به نظرم نسبت به بازار قیمت AQPvIPy3xH57cG uEElCjdr8pPQQ wIx91h2 x158mMTa2FVv... منطقیه 💸"
  ]
};